move urgency stuff to a module
